requirements.txt
nanome==0.34.7
biopython==1.79
